# Add minor performance bump to shredding