#### Problem

When finalizing a data shred the shredder manually constructs a cursor over an empty shred buffer. This is slightly slower than just using Vec's write implementation.

#### Summary of Changes

Used Vec's write implementation that just calls `extend_from_slice` internally. 
Added benches for shredding and deshredding

```
New baseline on master without signing
1Mb (8 * 1000 * 1000)
test bench_shredder   ... bench:  17,840,548 ns/iter (+/- 785,096)
test bench_shredder   ... bench:  17,823,871 ns/iter (+/- 780,049)
test bench_shredder   ... bench:  17,715,145 ns/iter (+/- 732,939)

1Mb (8 * 1000 * 1000) this PR and without signing
test bench_shredder   ... bench:  16,590,388 ns/iter (+/- 690,641)
test bench_shredder   ... bench:  16,818,323 ns/iter (+/- 554,646)
test bench_shredder   ... bench:  16,868,505 ns/iter (+/- 575,900)
```